Updated project to load all saved serialized scores from file when game starts.
Added Read() and ParseFileContents() to TextFileHandler.cs to read all file contents and return it as a list of strings.  Added LoadScores() to ScoreSaveSystemScript.cs to load all serialized scores into memory on game start as a list of deserialized ScoreEntry instances.  Also updated LoadScore() in ScoreSaveSystem to loop through the saved_score_entry_list to the next oldest ScoreEntry isntance.